Separate prime sieve into its own compilation unit
This way it is not depending on any other symengine classes.
