+++ symengine/tests/ntheory/test_ntheory.cpp
@@ -2,6 +2,7 @@
 #include <chrono>
 
 #include <symengine/ntheory.h>
+#include <symengine/prime_sieve.h>
 #include <symengine/rational.h>
 #include <symengine/add.h>
 #include <symengine/mul.h>
